Fixed fingerprinting time zone calculation at line #4373.
No, nothing, it was just wrong 'cos I copied it from some snippet that probably was meant to DETECT the current value of the DST, not the "clean" time zone (unaffected by the DST). Also, DST confuses me and I detest it, overall, which is probably why I didn't care enough about it. Sowwy. :)